[feat,to] 1. 导航导到泉水，而且血量不足80%则底盘断电 2. 双板通信频率再降低到12ms一次，id增大到电机之后 3. 云台板发送的包大小进一步减小 4. followpid小调整，kp减小一点，再增加deadzone，可能能消抖？

+++ UserApp/robot/sentry_omni_gimbal/robot.h
@@ -48,9 +48,9 @@ typedef struct {
   float Rotate_speed;
   int16_t Spin_speed;
   float Yaw_motor_angle;
-  float Mode_switch;
-  float Control_mode;
-  float Pause_flag;
+  uint8_t Mode_switch;
+  uint8_t Control_mode;
+  uint8_t Pause_flag;
 } Send_Data_RC_NEW;
 #endif
